requirements.txt
asgiref==3.7.2
astroid==2.15.5
attrs==23.1.0
black==23.3.0
click==8.1.3
Django==4.2.2
django-extensions==3.2.3
djangorestframework==3.14.0
drf-spectacular==0.26.2
flake8==6.0.0
flake8-django==1.3
inflection==0.5.1
jsonschema==4.17.3
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
pycodestyle==2.10.0
pyflakes==3.0.1
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
ruff==0.0.272
sqlparse==0.4.4
uritemplate==4.1.1
wrapt==1.15.0
